Lazy load reports module
